extract TestContainersFeature
