+++ starter-core/src/main/java/io/micronaut/starter/feature/database/TestContainersFeature.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2017-2022 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.starter.feature.database;
+
+import io.micronaut.starter.feature.Feature;
+import io.micronaut.starter.feature.FeatureContext;
+
+public abstract class TestContainersFeature implements Feature  {
+    private final TestContainers testContainers;
+
+    protected TestContainersFeature() {
+        this.testContainers = null;
+    }
+
+    protected TestContainersFeature(TestContainers testContainers) {
+        this.testContainers = testContainers;
+    }
+
+    @Override
+    public void processSelectedFeatures(FeatureContext featureContext) {
+        if (!featureContext.isPresent(TestContainers.class) && testContainers != null) {
+            featureContext.addFeature(testContainers);
+        }
+    }
+}
